+++ source/Knyaz.Optimus.Tests/EngineTests/JQueryFormTests.cs
@@ -1,5 +1,7 @@
 ﻿using NUnit.Framework;
 using Knyaz.Optimus.Tests.Resources;
+using Knyaz.Optimus.Tests.TestingTools;
+using Knyaz.Optimus.Tests.Tools;
 
 namespace Knyaz.Optimus.Tests.EngineTests
 {
@@ -9,7 +11,7 @@ namespace Knyaz.Optimus.Tests.EngineTests
 		[Test]
 		public void Smoke()
 		{
-			var engine = new Engine();
+			var engine = TestingEngine.BuildJint();
 			engine.ScriptExecutor.OnException += exception => System.Console.WriteLine(exception);
 			engine.Load("<html><head><script> " + R.JQueryJs + " </script>" +
 				"<script>" + R.JQueryFormJs + "</script></head><body></body></html>");