+++ source/Knyaz.Optimus.Tests/EngineTests/RequirejsTests.cs
@@ -2,6 +2,7 @@
 using Knyaz.Optimus.ResourceProviders;
 using NUnit.Framework;
 using Knyaz.Optimus.Tests.Resources;
+using Knyaz.Optimus.Tests.TestingTools;
 using Moq;
 
 namespace Knyaz.Optimus.Tests.EngineTests
@@ -20,7 +21,7 @@ namespace Knyaz.Optimus.Tests.EngineTests
 		[Test]
 		public void Smoke()
 		{
-			var engine = new Engine();
+			var engine = TestingEngine.BuildJint();
 			engine.ScriptExecutor.OnException += exception => System.Console.WriteLine(exception);
 			engine.Load("<html><head><script>\r\n" + requireJs + "\r\n</script></head><body></body></html>");
 		}
@@ -56,7 +57,7 @@ namespace Knyaz.Optimus.Tests.EngineTests
 		
 		private static List<string> OpenUrlAndGetLog(string url, IResourceProvider resourceProvider)
 		{
-			var engine = new Engine(resourceProvider);
+			var engine = TestingEngine.BuildJint(resourceProvider);
 			var log = new List<string>();
 			engine.Console.OnLog += o =>
 			{