+++ source/Knyaz.Optimus.Tests/UnitTests/Dom/Css/ComputedStyleTests.cs
@@ -3,6 +3,7 @@ using System.Linq;
 using System.Text;
 using Knyaz.Optimus.Dom.Elements;
 using Knyaz.Optimus.TestingTools;
+using Knyaz.Optimus.Tests.TestingTools;
 using NUnit.Framework;
 
 namespace Knyaz.Optimus.Tests.Dom.Css
@@ -12,7 +13,7 @@ namespace Knyaz.Optimus.Tests.Dom.Css
 	{
 		private Engine Load(string html)
 		{
-			var engine = new Engine { ComputedStylesEnabled = true };
+			var engine = TestingEngine.BuildJintCss();
 			engine.Load(new MemoryStream(Encoding.UTF8.GetBytes(html)));
 			return engine;
 		}