+++ source/Knyaz.Optimus/ScriptExecuting/Jint/EngineBuilderExtension.cs
@@ -5,7 +5,7 @@ namespace Knyaz.Optimus.ScriptExecuting.Jint
     public static class EngineBuilderExtension
     {
         /// <summary> Setup JINT javascript engine. </summary>
-        public static EngineBuilder UseJint(this EngineBuilder builder) => 
+        public static EngineBuilder UseJint(this EngineBuilder builder) =>
             builder.JsScriptExecutor(ctx => new JintJsScriptExecutor(ctx.Window, ctx.CreateXhr));
     }
 }