+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Optimus Changelog
 All notable changes to this project will be documented in this file.
 
+## [3.0.0]
+
+### Breaking changes
+- Removed Engine(IResourceProvider) constructor. Use EngineBuilder.
+- Removed Engine.OnWindowOpen event. Use EngineBuilder to configure Window object.
+
+
 ## [2.4]
 
 ### New features